Following trackers are checked in. EDK776	DEVICE_PATH_MESSAGING_SAS GUID Definition is not compliant with UEFI Spec EDK774	EBC X64 match ARightShift64() error EDK779	Network - EFI PXE referral doesn't work. EDK775	Update GlueLib to align with EdkII MDE library
git-svn-id: https://svn.code.sf.net/p/efidevkit/code/trunk/Edk@200 d4c01ab0-1994-4ed3-9bc5-3ada14f9202f

+++ Foundation/Library/EdkIIGlueLib/Library/BaseIoLibIntrinsic/IoHighLevel.c
@@ -56,7 +56,7 @@ IoOr8 (
   IN      UINT8                     OrData
   )
 {
-  return IoWrite8 (Port, IoRead8 (Port) | OrData);
+  return IoWrite8 (Port, (UINT8) (IoRead8 (Port) | OrData));
 }
 
 /**
@@ -84,7 +84,7 @@ IoAnd8 (
   IN      UINT8                     AndData
   )
 {
-  return IoWrite8 (Port, IoRead8 (Port) & AndData);
+  return IoWrite8 (Port, (UINT8) (IoRead8 (Port) & AndData));
 }
 
 /**
@@ -115,7 +115,7 @@ IoAndThenOr8 (
   IN      UINT8                     OrData
   )
 {
-  return IoWrite8 (Port, (IoRead8 (Port) & AndData) | OrData);
+  return IoWrite8 (Port, (UINT8) ((IoRead8 (Port) & AndData) | OrData));
 }
 
 /**
@@ -336,7 +336,7 @@ IoOr16 (
   IN      UINT16                    OrData
   )
 {
-  return IoWrite16 (Port, IoRead16 (Port) | OrData);
+  return IoWrite16 (Port, (UINT16) (IoRead16 (Port) | OrData));
 }
 
 /**
@@ -364,7 +364,7 @@ IoAnd16 (
   IN      UINT16                    AndData
   )
 {
-  return IoWrite16 (Port, IoRead16 (Port) & AndData);
+  return IoWrite16 (Port, (UINT16) (IoRead16 (Port) & AndData));
 }
 
 /**
@@ -395,7 +395,7 @@ IoAndThenOr16 (
   IN      UINT16                    OrData
   )
 {
-  return IoWrite16 (Port, (IoRead16 (Port) & AndData) | OrData);
+  return IoWrite16 (Port, (UINT16) ((IoRead16 (Port) & AndData) | OrData));
 }
 
 /**
@@ -1176,7 +1176,7 @@ MmioOr8 (
   IN      UINT8                     OrData
   )
 {
-  return MmioWrite8 (Address, MmioRead8 (Address) | OrData);
+  return MmioWrite8 (Address, (UINT8) (MmioRead8 (Address) | OrData));
 }
 
 /**
@@ -1204,7 +1204,7 @@ MmioAnd8 (
   IN      UINT8                     AndData
   )
 {
-  return MmioWrite8 (Address, MmioRead8 (Address) & AndData);
+  return MmioWrite8 (Address, (UINT8) (MmioRead8 (Address) & AndData));
 }
 
 /**
@@ -1236,7 +1236,7 @@ MmioAndThenOr8 (
   IN      UINT8                     OrData
   )
 {
-  return MmioWrite8 (Address, (MmioRead8 (Address) & AndData) | OrData);
+  return MmioWrite8 (Address, (UINT8) ((MmioRead8 (Address) & AndData) | OrData));
 }
 
 /**
@@ -1458,7 +1458,7 @@ MmioOr16 (
   IN      UINT16                    OrData
   )
 {
-  return MmioWrite16 (Address, MmioRead16 (Address) | OrData);
+  return MmioWrite16 (Address, (UINT16) (MmioRead16 (Address) | OrData));
 }
 
 /**
@@ -1486,7 +1486,7 @@ MmioAnd16 (
   IN      UINT16                    AndData
   )
 {
-  return MmioWrite16 (Address, MmioRead16 (Address) & AndData);
+  return MmioWrite16 (Address, (UINT16) (MmioRead16 (Address) & AndData));
 }
 
 /**
@@ -1518,7 +1518,7 @@ MmioAndThenOr16 (
   IN      UINT16                    OrData
   )
 {
-  return MmioWrite16 (Address, (MmioRead16 (Address) & AndData) | OrData);
+  return MmioWrite16 (Address, (UINT16) ((MmioRead16 (Address) & AndData) | OrData));
 }
 
 /**
